refactor: 提取 ASR 服务接口到 shared-types，消除跨包重复代码
将 ASRServiceEvents 和 IASRService 接口定义移动到 packages/shared-types/src/services/ 目录，
apps/backend 和 packages/esp32 通过 re-export 使用共享接口。

变更内容：
- 新增 packages/shared-types/src/services/asr.interface.ts（共享接口定义）
- 新增 packages/shared-types/src/services/index.ts（导出索引）
- 更新 packages/shared-types/src/index.ts（添加 services 导出）
- 更新 packages/shared-types/package.json（添加 ./services 导出路径）
- 更新 packages/shared-types/tsup.config.ts（添加 services 入口点）
- 更新 apps/backend/services/asr.interface.ts（re-export 共享接口）
- 更新 packages/esp32/src/services/asr.interface.ts（re-export 共享接口）

修复 issue #3178 中检测到的约 98 行重复代码。
