+++ packages/shared-types/src/services/index.ts
@@ -1,0 +1,10 @@
+/**
+ * 服务相关类型定义
+ */
+
+// ASR 服务接口
+export type {
+  ASRServiceEvents,
+  BaseASRServiceOptions,
+  IASRService,
+} from "./asr.interface";
